Make rustc compile without LLVM
Doesn't yet compile when `llvm-enabled = false` due to crate metadata loading needing LLVM
