+++ src/librustc_driver/lib.rs
@@ -48,6 +48,7 @@ extern crate rustc_metadata;
 extern crate rustc_mir;
 extern crate rustc_resolve;
 extern crate rustc_save_analysis;
+#[cfg(feature="llvm")]
 extern crate rustc_trans;
 extern crate rustc_typeck;
 extern crate serialize;
@@ -63,7 +64,9 @@ use pretty::{PpMode, UserIdentifiedItem};
 use rustc_resolve as resolve;
 use rustc_save_analysis as save;
 use rustc_save_analysis::DumpHandler;
+#[cfg(feature="llvm")]
 use rustc_trans::back::link;
+#[cfg(feature="llvm")]
 use rustc_trans::back::write::{RELOC_MODEL_ARGS, CODE_GEN_MODEL_ARGS};
 use rustc::dep_graph::DepGraph;
 use rustc::session::{self, config, Session, build_session, CompileResult};
@@ -175,6 +178,7 @@ pub fn run_compiler<'a>(args: &[String],
     let (sopts, cfg) = config::build_session_options_and_crate_config(&matches);
 
     if sopts.debugging_opts.debug_llvm {
+        #[cfg(feature="llvm")]
         rustc_trans::enable_llvm_debug();
     }
 
@@ -204,6 +208,7 @@ pub fn run_compiler<'a>(args: &[String],
     let mut sess = session::build_session_with_codemap(
         sopts, &dep_graph, input_file_path, descriptions, cstore.clone(), codemap, emitter_dest,
     );
+    #[cfg(feature="llvm")]
     rustc_trans::init(&sess);
     rustc_lint::register_builtins(&mut sess.lint_store.borrow_mut(), Some(&sess));
 
@@ -409,6 +414,7 @@ impl<'a> CompilerCalls<'a> for RustcDefaultCalls {
                     None,
                     descriptions.clone(),
                     cstore.clone());
+                #[cfg(feature="llvm")]
                 rustc_trans::init(&sess);
                 rustc_lint::register_builtins(&mut sess.lint_store.borrow_mut(), Some(&sess));
                 let mut cfg = config::build_configuration(&sess, cfg.clone());
@@ -607,7 +613,7 @@ impl RustcDefaultCalls {
                     };
                     let attrs = attrs.as_ref().unwrap();
                     let t_outputs = driver::build_output_filenames(input, odir, ofile, attrs, sess);
-                    let id = link::find_crate_name(Some(sess), attrs, input);
+                    let id: String = link::find_crate_name(Some(sess), attrs, input);
                     if *req == PrintRequest::CrateName {
                         println!("{}", id);
                         continue;
@@ -662,20 +668,36 @@ impl RustcDefaultCalls {
                     }
                 }
                 PrintRequest::RelocationModels => {
-                    println!("Available relocation models:");
-                    for &(name, _) in RELOC_MODEL_ARGS.iter() {
-                        println!("    {}", name);
+                    #[cfg(not(feature="llvm"))]
+                    panic!("LLVM is not enabled for this rustc version");
+
+                    #[cfg(feature="llvm")]
+                    {
+                        println!("Available relocation models:");
+                        for &(name, _) in RELOC_MODEL_ARGS.iter() {
+                            println!("    {}", name);
+                        }
+                        println!("");
                     }
-                    println!("");
                 }
                 PrintRequest::CodeModels => {
-                    println!("Available code models:");
-                    for &(name, _) in CODE_GEN_MODEL_ARGS.iter(){
-                        println!("    {}", name);
+                    #[cfg(not(feature="llvm"))]
+                    panic!("LLVM is not enabled for this rustc version");
+
+                    #[cfg(feature="llvm")]
+                    {
+                        println!("Available code models:");
+                        for &(name, _) in CODE_GEN_MODEL_ARGS.iter(){
+                            println!("    {}", name);
+                        }
+                        println!("");
                     }
-                    println!("");
                 }
                 PrintRequest::TargetCPUs | PrintRequest::TargetFeatures => {
+                    #[cfg(not(feature="llvm"))]
+                    panic!("LLVM is not enabled for this rustc version");
+
+                    #[cfg(feature="llvm")]
                     rustc_trans::print(*req, sess);
                 }
             }
@@ -715,6 +737,7 @@ pub fn version(binary: &str, matches: &getopts::Matches) {
         println!("commit-date: {}", unw(commit_date_str()));
         println!("host: {}", config::host_triple());
         println!("release: {}", unw(release_str()));
+        #[cfg(feature="llvm")]
         rustc_trans::print_version();
     }
 }
@@ -1008,6 +1031,10 @@ pub fn handle_options(args: &[String]) -> Option<getopts::Matches> {
     }
 
     if cg_flags.contains(&"passes=list".to_string()) {
+        #[cfg(not(feature="llvm"))]
+        panic!("LLVM is not enabled for this rustc version");
+
+        #[cfg(feature="llvm")]
         rustc_trans::print_passes();
         return None;
     }
@@ -1135,6 +1162,7 @@ pub fn diagnostics_registry() -> errors::registry::Registry {
     all_errors.extend_from_slice(&rustc_borrowck::DIAGNOSTICS);
     all_errors.extend_from_slice(&rustc_resolve::DIAGNOSTICS);
     all_errors.extend_from_slice(&rustc_privacy::DIAGNOSTICS);
+    #[cfg(feature="llvm")]
     all_errors.extend_from_slice(&rustc_trans::DIAGNOSTICS);
     all_errors.extend_from_slice(&rustc_const_eval::DIAGNOSTICS);
     all_errors.extend_from_slice(&rustc_metadata::DIAGNOSTICS);
@@ -1158,4 +1186,10 @@ pub fn main() {
                                      None,
                                      None));
     process::exit(result as i32);
+}
+
+
+#[cfg(not(feature="llvm"))]
+mod link {
+    include!("../librustc_trans/back/no_llvm_link.rs");
 }
